% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojoverse.R
\docType{package}
\name{ojoverse-package}
\alias{ojoverse}
\alias{ojoverse-package}
\title{ojoverse: All the Open Justice Oklahoma R Tools in One Place}
\description{
This package is a collection of all the Open Justice Oklahoma R tools. It is meant to be used for interactive workflows, not as a dependency for other packages. Fully inspired, and mostly written, by developers of the [{tidyverse}](https://github.com/tidyverse/tidyverse) package.
}
\author{
\strong{Maintainer}: Brancen Gregory \email{brancengregory@gmail.com}

}
\keyword{internal}
