% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{ojoverse_conflicts}
\alias{ojoverse_conflicts}
\title{Conflicts between the ojoverse and other packages}
\usage{
ojoverse_conflicts(only = NULL)
}
\arguments{
\item{only}{Set this to a character vector to restrict to conflicts only
with these packages.}
}
\description{
This function lists all the conflicts between packages in the ojoverse
and other packages that you have loaded.
}
\examples{
ojoverse_conflicts()

}
