% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{ojoverse_deps}
\alias{ojoverse_deps}
\title{List all ojoverse dependencies}
\usage{
ojoverse_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
ojoverse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOption("repos")}.}
}
\description{
List all ojoverse dependencies
}
