% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojoverse.R
\docType{package}
\name{ojoverse-package}
\alias{ojoverse}
\alias{ojoverse-package}
\title{ojoverse: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Brancen Gregory \email{brancengregory@gmail.com}

}
\keyword{internal}
